Do "Contract is mined" check before API sign private tx spec scenarios
Unlike the other similar tests, the deploy context step returns only the tx hash.  The steps in the "Contract is mined" scenario retrieve the tx receipt and so create the Contract object used in the later scenarios.  Performing these steps before each scenario ensure a Contract object is available for testing.