Fix bug that causes tests for externally signed txs to fail
The from address of the externally signed tx was being compared against the default account of the node originally used to send the tx, as opposed to the external wallet account used to sign the tx.  When running these tests locally against a quorum-examples 7nodes network the tests passed, but these were false positives.  This was because external wallet 1 used in these tests is the same as node 1's default account.

This fix correctly compares the from address of the tx to the external wallet address or node's default address where appropriate.

To prevent similar issues from occurring in the future, the external wallets have been replaced with newly generated wallets not found in the quorum-examples 7nodes project.  Both are protected with empty passwords.