run tests using Docker (#60)
The workflow run.yml is bloated with tools and env setup which duplicates those in Dockerfile.
This is to remove the duplication and centralize the updates only in Dockerfile by building Docker image prior running the suite. Docker image is built once and cached to improve the total execution time.