Improve PE-enabled network upgrade test stability (#68)
While running the network upgrade scenario using raft the quorum process panics when it gets into a BadBlock state (caused by receiving a privacy enhanced transaction while privacy enhancements are disabled).
The tests then check that the state of the quorum container is down.

When running on AWS it seems there is some delay between the quorum process dying and the state of the docker container being reported as down. The test was updated to wait for the container to reach the state rather than just check current container state.